+++ dependencies/maven/artifacts.bzl
@@ -1,0 +1,7 @@
+artifacts = [
+    "commons-io:commons-io",
+    "com.univocity:univocity-parsers",
+    "org.sharegov:mjson",
+    "com.google.code.gson:gson",
+    "javax.xml.stream:stax-api",
+]
