+++ dependencies/maven/artifacts.snapshot
@@ -1,0 +1,134 @@
+@maven//:ch_qos_logback_logback_classic
+@maven//:ch_qos_logback_logback_classic_1_2_3
+@maven//:ch_qos_logback_logback_core
+@maven//:ch_qos_logback_logback_core_1_2_3
+@maven//:com_google_android_annotations
+@maven//:com_google_android_annotations_4_1_1_4
+@maven//:com_google_api_grpc_proto_google_common_protos
+@maven//:com_google_api_grpc_proto_google_common_protos_1_12_0
+@maven//:com_google_code_findbugs_annotations
+@maven//:com_google_code_findbugs_annotations_3_0_1
+@maven//:com_google_code_findbugs_jsr305
+@maven//:com_google_code_findbugs_jsr305_2_0_2
+@maven//:com_google_code_gson_gson
+@maven//:com_google_code_gson_gson_2_8_5
+@maven//:com_google_errorprone_error_prone_annotations
+@maven//:com_google_errorprone_error_prone_annotations_2_3_2
+@maven//:com_google_guava_guava
+@maven//:com_google_guava_guava_23_0
+@maven//:com_google_j2objc_j2objc_annotations
+@maven//:com_google_j2objc_j2objc_annotations_1_1
+@maven//:com_google_protobuf_protobuf_java
+@maven//:com_google_protobuf_protobuf_java_3_5_1
+@maven//:com_univocity_univocity_parsers
+@maven//:com_univocity_univocity_parsers_2_8_1
+@maven//:commons_io_commons_io
+@maven//:commons_io_commons_io_2_3
+@maven//:io_cucumber_cucumber_core
+@maven//:io_cucumber_cucumber_core_5_1_3
+@maven//:io_cucumber_cucumber_expressions
+@maven//:io_cucumber_cucumber_expressions_8_3_1
+@maven//:io_cucumber_cucumber_gherkin
+@maven//:io_cucumber_cucumber_gherkin_5_1_3
+@maven//:io_cucumber_cucumber_gherkin_vintage
+@maven//:io_cucumber_cucumber_gherkin_vintage_5_1_3
+@maven//:io_cucumber_cucumber_java
+@maven//:io_cucumber_cucumber_java_5_1_3
+@maven//:io_cucumber_cucumber_junit
+@maven//:io_cucumber_cucumber_junit_5_1_3
+@maven//:io_cucumber_cucumber_plugin
+@maven//:io_cucumber_cucumber_plugin_5_1_3
+@maven//:io_cucumber_datatable
+@maven//:io_cucumber_datatable_3_2_1
+@maven//:io_cucumber_docstring
+@maven//:io_cucumber_docstring_5_1_3
+@maven//:io_cucumber_tag_expressions
+@maven//:io_cucumber_tag_expressions_2_0_4
+@maven//:io_grpc_grpc_api
+@maven//:io_grpc_grpc_api_1_24_1
+@maven//:io_grpc_grpc_context
+@maven//:io_grpc_grpc_context_1_24_1
+@maven//:io_grpc_grpc_core
+@maven//:io_grpc_grpc_core_1_24_1
+@maven//:io_grpc_grpc_netty
+@maven//:io_grpc_grpc_netty_1_24_1
+@maven//:io_grpc_grpc_protobuf
+@maven//:io_grpc_grpc_protobuf_1_24_1
+@maven//:io_grpc_grpc_protobuf_lite
+@maven//:io_grpc_grpc_protobuf_lite_1_24_1
+@maven//:io_grpc_grpc_stub
+@maven//:io_grpc_grpc_stub_1_24_1
+@maven//:io_grpc_grpc_testing
+@maven//:io_grpc_grpc_testing_1_24_1
+@maven//:io_netty_netty_all
+@maven//:io_netty_netty_all_4_1_50_Final
+@maven//:io_netty_netty_buffer
+@maven//:io_netty_netty_buffer_4_1_50_Final
+@maven//:io_netty_netty_codec
+@maven//:io_netty_netty_codec_4_1_50_Final
+@maven//:io_netty_netty_codec_http
+@maven//:io_netty_netty_codec_http2
+@maven//:io_netty_netty_codec_http2_4_1_50_Final
+@maven//:io_netty_netty_codec_http_4_1_50_Final
+@maven//:io_netty_netty_codec_socks
+@maven//:io_netty_netty_codec_socks_4_1_50_Final
+@maven//:io_netty_netty_common
+@maven//:io_netty_netty_common_4_1_50_Final
+@maven//:io_netty_netty_handler
+@maven//:io_netty_netty_handler_4_1_50_Final
+@maven//:io_netty_netty_handler_proxy
+@maven//:io_netty_netty_handler_proxy_4_1_50_Final
+@maven//:io_netty_netty_resolver
+@maven//:io_netty_netty_resolver_4_1_50_Final
+@maven//:io_netty_netty_tcnative_boringssl_static
+@maven//:io_netty_netty_tcnative_boringssl_static_2_0_25_Final
+@maven//:io_netty_netty_transport
+@maven//:io_netty_netty_transport_4_1_50_Final
+@maven//:io_opencensus_opencensus_api
+@maven//:io_opencensus_opencensus_api_0_21_0
+@maven//:io_opencensus_opencensus_contrib_grpc_metrics
+@maven//:io_opencensus_opencensus_contrib_grpc_metrics_0_21_0
+@maven//:io_perfmark_perfmark_api
+@maven//:io_perfmark_perfmark_api_0_17_0
+@maven//:javax_annotation_javax_annotation_api
+@maven//:javax_annotation_javax_annotation_api_1_3_2
+@maven//:javax_xml_stream_stax_api
+@maven//:javax_xml_stream_stax_api_1_0_2
+@maven//:junit_junit
+@maven//:junit_junit_4_12
+@maven//:net_bytebuddy_byte_buddy
+@maven//:net_bytebuddy_byte_buddy_1_6_4
+@maven//:net_bytebuddy_byte_buddy_agent
+@maven//:net_bytebuddy_byte_buddy_agent_1_6_4
+@maven//:net_jcip_jcip_annotations
+@maven//:net_jcip_jcip_annotations_1_0
+@maven//:org_antlr_antlr4_runtime
+@maven//:org_antlr_antlr4_runtime_4_7_1
+@maven//:org_apiguardian_apiguardian_api
+@maven//:org_apiguardian_apiguardian_api_1_1_0
+@maven//:org_codehaus_mojo_animal_sniffer_annotations
+@maven//:org_codehaus_mojo_animal_sniffer_annotations_1_17
+@maven//:org_hamcrest_hamcrest
+@maven//:org_hamcrest_hamcrest_2_2
+@maven//:org_hamcrest_hamcrest_all
+@maven//:org_hamcrest_hamcrest_all_1_3
+@maven//:org_hamcrest_hamcrest_core
+@maven//:org_hamcrest_hamcrest_core_1_3
+@maven//:org_hamcrest_hamcrest_library
+@maven//:org_hamcrest_hamcrest_library_1_3
+@maven//:org_mockito_mockito_core
+@maven//:org_mockito_mockito_core_2_6_4
+@maven//:org_objenesis_objenesis
+@maven//:org_objenesis_objenesis_2_5
+@maven//:org_sharegov_mjson
+@maven//:org_sharegov_mjson_1_4_1
+@maven//:org_slf4j_jcl_over_slf4j
+@maven//:org_slf4j_jcl_over_slf4j_1_7_20
+@maven//:org_slf4j_log4j_over_slf4j
+@maven//:org_slf4j_log4j_over_slf4j_1_7_20
+@maven//:org_slf4j_slf4j_api
+@maven//:org_slf4j_slf4j_api_1_7_28
+@maven//:org_slf4j_slf4j_simple
+@maven//:org_slf4j_slf4j_simple_1_7_20
+@maven//:org_zeroturnaround_zt_exec
+@maven//:org_zeroturnaround_zt_exec_1_10
