--- dependencies/maven/artifacts/com/google/code/gson/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "gson",
-    exports = [
-        "//external:jar/com/google/code/gson/gson"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
