--- dependencies/maven/artifacts/com/univocity/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "univocity-parsers",
-    exports = [
-        "//external:jar/com/univocity/univocity-parsers"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
