--- dependencies/maven/artifacts/commons-io/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "commons-io",
-    exports = [
-        "//external:jar/commons-io/commons-io"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
