--- dependencies/maven/artifacts/javax/xml/stream/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "stax-api",
-    exports = [
-        "//external:jar/javax/xml/stream/stax-api"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
