--- dependencies/maven/artifacts/junit/BUILD
@@ -1,14 +1,0 @@
-java_library(
-    name = "junit",
-    exports = [
-        "//external:jar/junit/junit"
-    ],
-    runtime_deps = [
-        "//dependencies/maven/artifacts/org/hamcrest:hamcrest-core"
-    ],
-    visibility = [
-        "//dependencies/maven/artifacts:__subpackages__"
-    ]
-)
-
-
