--- dependencies/maven/artifacts/org/hamcrest/BUILD
@@ -1,11 +1,0 @@
-java_library(
-    name = "hamcrest-core",
-    exports = [
-        "//external:jar/org/hamcrest/hamcrest-core"
-    ],
-    visibility = [
-        "//dependencies/maven/artifacts:__subpackages__"
-    ]
-)
-
-
