--- dependencies/maven/artifacts/org/sharegov/BUILD
@@ -1,14 +1,0 @@
-java_library(
-    name = "mjson",
-    exports = [
-        "//external:jar/org/sharegov/mjson"
-    ],
-    runtime_deps = [
-        "//dependencies/maven/artifacts/junit:junit"
-    ],
-    visibility = [
-        "//visibility:public"
-    ]
-)
-
-
