--- dependencies/maven/dependencies.yaml
@@ -1,34 +1,0 @@
-options:
-  languages: ["java"]
-  resolverType: coursier
-  resolvers:
-  - id: mavencentral
-    type: default
-    url: https://repo.maven.apache.org/maven2/
-  strictVisibility: true
-  transitivity: runtime_deps
-  versionConflictPolicy: highest
-  thirdPartyDirectory: dependencies/maven/artifacts
-
-
-dependencies:
-  commons-io:
-    commons-io:
-      version: "2.6"
-      lang: java
-  com.univocity:
-    univocity-parsers:
-      version: "2.8.1"
-      lang: java
-  org.sharegov:
-    mjson:
-      version: "1.4.1"
-      lang: java
-  com.google.code.gson:
-    gson:
-      version: "2.8.5"
-      lang: java
-  javax.xml.stream:
-    stax-api:
-      version: "1.0-2"
-      lang: java
