+++ test/query/BUILD
@@ -1,35 +1,32 @@
+load("@graknlabs_common//test/server:rules.bzl", "grakn_test")
+
 #########################
 # Graql Java Query Test #
 #########################
 
-java_test(
+grakn_test(
     name = "graql-java",
     test_class = "grakn.doc.test.query.GraqlJavaTest",
     srcs = [":graql-java-tests"],
     deps = [
         "@graknlabs_client_java//:client-java",
         "@graknlabs_graql//java:graql",
-        "@graknlabs_grakn_core//server:server",
-        "@graknlabs_grakn_core//test/rule:grakn-test-server"
-    ],
-    classpath_resources = [
-        "@graknlabs_grakn_core//test/resources:logback-test",
     ],
     data = [
         "//files:social-network-kg", # filegroup defined in files/BUILD
         "//files:phone-calls-kg", # filegroup defined in files/BUILD
-        "//test/conf:cassandra-embedded.yaml",
-        "@graknlabs_grakn_core//server:conf/grakn.properties"
-    ]
+    ],
+    grakn_artifact = "@graknlabs_grakn_core_artifact//file",
 )
 
 genrule(
     name = "graql-java-tests",
     srcs = [
         "graql-java-tests.py",
+        "templates/GraqlJavaTest.java",
         "//:content" # filegroup defined in root/BUILD
     ],
-    cmd = "python3 $(location graql-java-tests.py) $@ $(locations //:content)",
+    cmd = "python3 $(location graql-java-tests.py) $(location templates/GraqlJavaTest.java) $@ $(locations //:content)",
     outs = ["GraqlJavaTest.java"]
 )
 
@@ -37,33 +34,28 @@ genrule(
 # Graql Lang Query Test #
 #########################
 
-java_test(
+grakn_test(
     name = "graql-lang",
     test_class = "grakn.doc.test.query.GraqlLangTest",
     srcs = [":graql-lang-tests"],
     deps = [
         "@graknlabs_client_java//:client-java",
         "@graknlabs_graql//java:graql",
-        "@graknlabs_grakn_core//server:server",
-        "@graknlabs_grakn_core//test/rule:grakn-test-server"
-    ],
-    classpath_resources = [
-        "@graknlabs_grakn_core//test/resources:logback-test",
     ],
     data = [
         "//files:social-network-kg", # filegroup defined in files/BUILD
         "//files:phone-calls-kg", # filegroup defined in files/BUILD
-        "//test/conf:cassandra-embedded.yaml",
-        "@graknlabs_grakn_core//server:conf/grakn.properties"
-    ]
+    ],
+    grakn_artifact = "@graknlabs_grakn_core_artifact//file",
 )
 
 genrule(
     name = "graql-lang-tests",
     srcs = [
         "graql-lang-tests.py",
+        "templates/GraqlLangTest.java",
         "//:content" # filegroup defined in root/BUILD
     ],
-    cmd = "python3 $(location graql-lang-tests.py) $@ $(locations //:content)",
+    cmd = "python3 $(location graql-lang-tests.py) $(location templates/GraqlLangTest.java) $@ $(locations //:content) ",
     outs = ["GraqlLangTest.java"]
 )
