+++ test/query/graql-lang-tests.py
@@ -2,75 +2,9 @@ import re
 import sys
 from io import open
 
-output_path, markdown_files = sys.argv[1], sys.argv[2:]
+graql_lang_test_template_path, output_path, markdown_files = sys.argv[1], sys.argv[2], sys.argv[3:]
 
-graql_lang_test_template = """
-package grakn.doc.test.query;
-
-import grakn.client.GraknClient;
-import grakn.core.test.rule.GraknTestServer;
-import graql.lang.Graql;
-import graql.lang.pattern.Pattern;
-import graql.lang.query.GraqlQuery;
-import org.junit.*;
-
-import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.stream.Stream;
-
-public class GraqlLangTest {
-    @ClassRule
-    public static final GraknTestServer server = new GraknTestServer(
-        Paths.get("external/graknlabs_grakn_core/server/conf/grakn.properties"),
-        Paths.get("test/conf/cassandra-embedded.yaml")
-    );
-
-    static GraknClient client;
-    static GraknClient.Session session;
-    GraknClient.Transaction transaction;
-
-
-    @BeforeClass
-    public static void loadSocialNetwork() {
-        client = new GraknClient(server.grpcUri().toString());
-        session = client.session("social_network");
-        GraknClient.Transaction transaction = session.transaction().write();
-
-        try {
-            byte[] encoded = Files.readAllBytes(Paths.get("files/social-network/schema.gql"));
-            String query = new String(encoded, StandardCharsets.UTF_8);
-            transaction.execute((GraqlQuery) Graql.parse(query)).get();
-
-            encoded = Files.readAllBytes(Paths.get("files/phone-calls/schema.gql"));
-            query = new String(encoded, StandardCharsets.UTF_8);
-            transaction.execute((GraqlQuery) Graql.parse(query)).get();
-
-            transaction.commit();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Before
-    public void openTransaction() {
-        transaction = session.transaction().write();
-    }
-
-    @After
-    public void closeTransaction() {
-        transaction.close();
-    }
-
-    @AfterClass
-    public static void closeSession() {
-        session.close();
-    }
-
-    // TEST METHODS PLACEHOLDER
-}
-"""
+graql_lang_test_template = open(graql_lang_test_template_path, 'r').read()
 
 graql_lang_test_method_template = """
     @Test
